Nhat done format input cozeAI and fixed delete exam java class, bank test java oop (java class).

+++ frontend-web-lecture/src/pages/bank-exams-java-oop/delete-test/ModalDeleteTestJavaOop.vue
@@ -2,6 +2,8 @@
 
 
 import BankTestJavaOopDao from "@/daos/BankTestJavaOopDao.js";
+import ExamDao from "@/daos/ExamDao.js";
+import ExamJavaOopDao from "@/daos/ExamJavaOopDao.js";
 
 export default {
   name: "ModalDeleteTestJavaOop",
@@ -22,13 +24,26 @@ export default {
     },
 
     async handleDeleteTest() {
-      //check them sv có làm
+      //check them sv có làm, ktra them bai kiem tra co duoc tao hay ko
       if(this.bankTestJavaOopIDToDelete) {
-        let status = BankTestJavaOopDao
-            .delete_Java_Test_Oop_By_BankTestJavaOopID(this.bankTestJavaOopIDToDelete)
-        if(status) {
-          alert("Delete test java oop successfully.");
-          window.location.reload();
+        let statusStudentAccess = await ExamDao
+            .getStatus_Access_Student_To_Exam_By_ExamID(this.examID);
+
+        let existExam = await ExamJavaOopDao.get_Exist_Exam_By_BankTestJavaOopID(Number(this.bankTestJavaOopIDToDelete));
+        console.log('Exist exam: ', existExam);
+        if(existExam) {
+          alert("Can't delete test java class because it is created to exam.");
+        } else {
+          if(!statusStudentAccess) {
+            let status = BankTestJavaOopDao
+                .delete_Java_Test_Oop_By_BankTestJavaOopID(this.bankTestJavaOopIDToDelete)
+            if(status) {
+              alert("Delete test java class successfully.");
+              window.location.reload();
+            }
+          } else {
+            alert("There is student is currently accessing the test. Please wait until finish.");
+          }
         }
       }
     }
